+++ packages/components/stories/GridList.stories.tsx
@@ -90,7 +90,7 @@ export const DragAndDrop: Story = {
 		});
 
 		const { dragAndDropHooks } = useDragAndDrop({
-			getItems: (keys) => [...keys].map((key) => ({ 'text/plain': list.getItem(key).name })),
+			getItems: (keys) => [...keys].map((key) => ({ 'text/plain': list.getItem(key)?.name ?? '' })),
 			onReorder(e) {
 				if (e.target.dropPosition === 'before') {
 					list.moveBefore(e.target.key, e.keys);