+++ packages/progress-bubbles/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @launchpad-ui/progress-bubbles
 
+## 0.7.36
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @launchpad-ui/popover@0.11.36
+
 ## 0.7.35
 
 ### Patch Changes